add might_sleep in rt_spin_lock_fastlock
Ingo,

Every so often we get bit by a bug "scheduling in atomic", and it comes
from a rtmutex spin_lock.  The bug only happens when that lock has
contention, so we miss it a lot.

This patch adds a might_sleep() to the rt_spin_lock_fastlock to find
bugs where we can schedule in atomic.

The one place that exists now is from do_page_fault and sending a
signal.  I wrote a simple crash program that segfaults (attached) and
with this patch, I get the warning.

-- Steve

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>