Enable HPET on ICH3 and ICH4
ICH3 and ICH4 have undocumented HPET capabilities. This patch enables HPET
for platforms based around these ICHs. 

Tested on various ICH3 and ICH4 platforms.

Because HPET is not officially documented for ICH3/4 and may not have been
validated by chipset folks, we're on thin ice here. I'd recommend testing 
this patch in -hrt or -mm for a while and wait for success/failure reports
before feeding it upstream.

Signed-off-by: Udo A. Steinberg <us15@os.inf.tu-dresden.de>