Don't call mcount from vsyscall_fn's
This bit me in the butt.

I couldn't understand why my init app was segfaulting, with a kernel
address, but a user RIP and RSP.  Well, the RIP I think was bogus, but
the kernel address was always the start of "mcount".  Looking deeper, I
printed out what was in the RSP (even though it was a user stack).  It
ended up showing me that the calling address was from the VDSO area.
Looking even further, I found the offending culprit, which was
vread_hpet.

Looking at the assembly dump, I saw the vread_hpet was calling mcount,
but I could not see it in the code. Nor could I see it in hpet.i (-E
option of compiling).

Well, I guess Ingo is a magician when it comes to compiler tricks, and
has the mcount being called by "every!!" function, unless you add the
"notrace" option.

This patch adds the notrace to vsyscall_fn, so that we don't have user
land apps calling mcount and crashing!

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>