kvm: move the apic timer migration
Move apic timer migration to a place where it does not cause the
"might sleep while atomic" check. The original place calls 
hrtimer_cancel in a preempt disabled region, which is fine in mainline,
but preempt-rt changes hrtimer_cancel, that the caller sleeps on a
wait_queue, when the callback of the timer is currently active.

Scheduled to go to mainline as well.

Signed-off-by: Avi Kiviti <avi@qumranet.com>
Signed-off-by: Thomas Gleixner <tgxl@linutronix.de>