RT: fix spinlock preemption feature when PREEMPT_RT is enabled
    [ The following text is in the "utf-8" character set. ]
    [ Your display is set for the "iso-8859-1" character set.  ]
    [ Some characters may be displayed incorrectly. ]

kernel/spinlock.c implements two versions of spinlock wrappers around
the arch-specific implementations:

1) A simple passthrough which implies disabled preemption while spinning

2) A "preemptible waiter" version which uses trylock.

Currently, PREEMPT && SMP will turn on the preemptible feature, and
lockdep or PREEMPT_RT will disable it.  Disabling the feature for
lockdep makes perfect sense, but PREEMPT_RT is counter-intuitive.  My
guess is that this was inadvertent, so this patch once again enables
the feature for PREEMPT_RT.

(Since PREEMPT is set for PREEMPT_RT, we simply get rid of the extra
condition).

I have tested the PREEMPT_RT kernel with this patch and all seems well.
Therefore, if there *is* an issue with running preemptible versions of
these spinlocks under PREEMPT_RT, it is not immediately apparent why. 

Signed-off-by: Gregory Haskins <ghaskins@novell.com>