Revert d04ea10ba1ea mmc: sdhci: don't provide hard irq handler
This workaround cannot work because the logic in the SDHCI driver is:

interrupt
  hard_interrupt_handler()
    wake_thread_handler()

thread_handler()
  poke_hardware()
  wait_for_interrupt()

Now the reverted commit changed that to:

interrupt
    wake_thread()

thread()
  hard_interrupt_handler()
    thread_handler()
    poke_hardware()
    wait_for_interrupt()

Doesn't really work well because thread handler waits for
hard_interrupt_handler() to wake it up again.....

This revert reintroduces the problem which was tried to be solved by
this patch. This can't be solved at the driver level, this needs a
solution at the irq core code.

Reported-by: Nathan Sullivan <nathan.sullivan@ni.com>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>