preempt-realtime-ia64
Hi,

This is a first version of my port of Ingo's -rt kernel to the IA64 arch.

So far the kernel boots with PREEMPT_RT enabled (on a 4-cpu tiger), and 
that's about it. I've not done extensive tests (only scripts/rt-tester), 
nor any measurements of any kind.

There's very probably many bugs I'm not aware of.

But there is already one thing I know should be fixed : I've changed the 
declaration of (struct zone).lock (in include/linux/mmzone.h) from 
spinlock_t to raw_spinlock_t. 

I did this because on IA64, cpu_idle(), which is not allowed to call 
schedule(), calls check_pgt_cache(). I guess this could be fixed by moving 
this call to another kernel thread... ideas are welcome.


	Simon.




Signed-off-by: Simon.Derr@bull.net

 arch/ia64/Kconfig                 |   64 +++++++++++++++++++++++++
 arch/ia64/kernel/asm-offsets.c    |    2 
 arch/ia64/kernel/entry.S          |   25 +++++-----
 arch/ia64/kernel/fsys.S           |   21 ++++++++
 arch/ia64/kernel/iosapic.c        |   34 ++++++++++++-
 arch/ia64/kernel/mca.c            |    2 
 arch/ia64/kernel/perfmon.c        |    6 +-
 arch/ia64/kernel/process.c        |   14 +++--
 arch/ia64/kernel/sal.c            |    2 
 arch/ia64/kernel/salinfo.c        |    6 +-
 arch/ia64/kernel/semaphore.c      |    8 +--
 arch/ia64/kernel/signal.c         |    8 +++
 arch/ia64/kernel/smp.c            |   16 ++++++
 arch/ia64/kernel/smpboot.c        |    3 +
 arch/ia64/kernel/time.c           |   74 +++++++++++++++++++----------
 arch/ia64/kernel/traps.c          |   10 ++--
 arch/ia64/kernel/unwind.c         |    4 -
 arch/ia64/kernel/unwind_i.h       |    2 
 arch/ia64/mm/init.c               |    2 
 arch/ia64/mm/tlb.c                |    2 
 include/asm-ia64/irqflags.h       |   95 ++++++++++++++++++++++++++++++++++++++
 include/asm-ia64/mmu_context.h    |    2 
 include/asm-ia64/percpu.h         |   21 +++++++-
 include/asm-ia64/processor.h      |    6 +-
 include/asm-ia64/rtc.h            |    7 ++
 include/asm-ia64/rwsem.h          |   32 ++++++------
 include/asm-ia64/sal.h            |    2 
 include/asm-ia64/semaphore.h      |   53 +++++++++++++--------
 include/asm-ia64/spinlock.h       |   26 ++++------
 include/asm-ia64/spinlock_types.h |    4 -
 include/asm-ia64/system.h         |   67 --------------------------
 include/asm-ia64/thread_info.h    |    1 
 include/asm-ia64/tlb.h            |   10 ++--
 33 files changed, 438 insertions(+), 193 deletions(-)